推荐功能
完成基于余弦相似度的推荐功能以及相关数据库存储

+++ src/main/java/com/hk/repository/ParagraphRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
  * smallHK
  * 2019/3/27 20:54
  */
-public interface ParagraphRepository extends CrudRepository<Paragraph, Integer> {
+public interface ParagraphRepository extends CrudRepository<Paragraph, Long> {
 
     List<Paragraph> findAllByChapterId(Integer chapterId);
